Refactor Parser class into ParserService
Parser class was renamed to ParserService and moved to the service package to better fit the architectural patterns. Dependency injection was implemented for MessageDigest and HttpHeaders for testability benefits and adherence to inversion of control principle. Improved code readability in methods by refactoring complex multilines into separate methods or simpler lines. Additional constants were introduced for elegance and to avoid magic strings and numbers.