Refactor VodRetriever to VodService and enhance dependency injection
Moved VodRetriever to service package and renamed to VodService for consistency in project structure. Introduced usage of dependency injection for the classes ParserService, TokenRequest, RestTemplate, and ConfigLoader to make the codebase cleaner and easier to unit test. Updated methods retrieveVod and recoverVod to return Map<String, Quality> instead of Feeds for clarity and improved functionality.