Integrate Guice for dependency injection
In the Main class, Guice was added to manage the dependency injection to create an instance of TwitchForgeBot. This change moved from manually instantiating VodRetriever and passing as an argument to TwitchForgeBot, to retrieving the instance of TwitchForgeBot by injecting the dependency via Guice, thus ensuring cleaner code and facilitating testing. Additionally, error handling was optimized from printing stack trace to logging the error message.